requirements.txt
faiss-cpu
groq
PyPDF2
langchain_google_genai
langchain
streamlit
langchain_community
python-dotenv
pypdf
langchain-openai